LUCENE-1027:  Added better formatting of doubles, added wikipedia-flush-by-RAM for comparison
git-svn-id: https://svn.apache.org/repos/asf/lucene/java/trunk@583792 13f79535-47bb-0310-9956-ffa450edef68

+++ contrib/benchmark/src/java/org/apache/lucene/benchmark/byTask/utils/Config.java
@@ -376,7 +376,7 @@ public class Config {
         else if (a instanceof double[]) {
           double ad[] = (double[]) a;
           int n = roundNum % ad.length;
-          sb.append(Format.format(2, (float) ad[n],template));
+          sb.append(Format.format(2, ad[n],template));
         }
         else {
           boolean ab[] = (boolean[]) a;
